Модуляция, импорт и экспорт
